requirements.txt
cellpose==3.0
matplotlib==3.7.2
numpy==1.24.3
numpy==1.24.4
opencv_python_headless==4.11.0.86
Pillow==11.3.0
PyQt5==5.15.11
PyQt5_sip==12.13.0
scikit_image==0.19.3
svgwrite==1.4.3
torch==2.4.1+cu118
torchvision==0.19.1+cu118
